// src/app/modules/school-admin/models/school-service-provider-config-view.ts
/**
 * Communication
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * School Service Configuration Data
 */
export interface SchoolServiceProviderConfigView { 
    /**
     * Gets or sets ServiceProviderConfigurationId
     */
    serviceProviderConfigurationId?: number;
    /**
     * Gets or sets Value
     */
    value?: string;
}