// src/app/modules/school-admin/models/service-code-view.ts
/**
 * Communication
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Gets or Sets ServiceCodeView
 */
export interface ServiceCodeView { 
    /**
     * Gets or Sets Code
     */
    code?: string;

    schoolId?: string;
    /**
     * Gets or Sets StatusFlag
     */
    statusFlag?: string;
}